projects/ai-agent/requirements.txt
# AI Agent Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
python-dotenv>=1.0.0
pydantic>=2.4.0

# Algorand Dependencies
algokit-utils>=2.3.0
py-algorand-sdk>=2.6.0

# AI/ML Dependencies  
google-generativeai>=0.3.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
